#4: implementation of list and validation of cqr's

+++ src/src/main/kotlin/com/akkarimzai/eventticket/models/event/ListEventQuery.kt
@@ -8,6 +8,6 @@ data class ListEventQuery(val title: String?,
     val from: LocalDateTime?,
     val to: LocalDateTime?,
     @Min(0)
-    val page: Long,
+    val page: Int,
     @Min(1)
-    val size: Long)
+    val size: Int)
